Fix ktor test app setup
